Add TransactionTestCase support for TenantTestCase and FastTenantTestCase
